Moved `no_data_page` into `page_template
