+++ src/plugins/kibana_react/public/page_template/index.ts
@@ -8,3 +8,4 @@
 
 export { KibanaPageTemplate, KibanaPageTemplateProps } from './page_template';
 export { KibanaPageTemplateSolutionNavAvatar } from './solution_nav';
+export * from './no_data_page';
