+++ src/plugins/kibana_react/public/page_template/no_data_page/index.ts
