+++ src/plugins/kibana_react/public/page_template/no_data_page/no_data_card/elastic_agent_card.test.tsx
@@ -10,8 +10,8 @@ import { shallow } from 'enzyme';
 import React from 'react';
 import { ElasticAgentCard } from './elastic_agent_card';
 
-jest.mock('../../context', () => ({
-  ...jest.requireActual('../../context'),
+jest.mock('../../../context', () => ({
+  ...jest.requireActual('../../../context'),
   useKibana: jest.fn().mockReturnValue({
     services: {
       http: { basePath: { prepend: jest.fn((path: string) => (path ? path : 'path')) } },
