+++ src/plugins/kibana_react/public/page_template/no_data_page/no_data_card/elastic_agent_card.tsx
@@ -12,26 +12,28 @@ import React, { FunctionComponent } from 'react';
 import { i18n } from '@kbn/i18n';
 import { CoreStart } from 'kibana/public';
 import { EuiButton, EuiCard } from '@elastic/eui';
-import { useKibana } from '../../context';
+import { useKibana } from '../../../context';
 import { NoDataPageActions, NO_DATA_RECOMMENDED } from '../no_data_page';
 
-export type ElasticBeatsCardProps = NoDataPageActions & {
+export type ElasticAgentCardProps = NoDataPageActions & {
   solution: string;
 };
 
-export const ElasticBeatsCard: FunctionComponent<ElasticBeatsCardProps> = ({
-  recommended,
-  href = 'app/home#/tutorial',
-  button,
+/**
+ * Applies extra styling to a typical EuiAvatar
+ */
+export const ElasticAgentCard: FunctionComponent<ElasticAgentCardProps> = ({
   solution,
+  recommended = true,
+  href = 'app/integrations/browse',
+  button,
   ...cardRest
 }) => {
   const {
-    services: { http, uiSettings },
+    services: { http },
   } = useKibana<CoreStart>();
   const addBasePath = http.basePath.prepend;
   const basePathUrl = '/plugins/kibanaReact/assets/';
-  const IS_DARK_THEME = uiSettings.get('theme:darkMode');
 
   const footer =
     typeof button !== 'string' && typeof button !== 'undefined' ? (
@@ -39,8 +41,8 @@ export const ElasticBeatsCard: FunctionComponent<ElasticBeatsCardProps> = ({
     ) : (
       <EuiButton href={href} onClick={cardRest?.onClick} target={cardRest?.target} fill>
         {button ||
-          i18n.translate('kibana-react.noDataPage.elasticBeatsCard.buttonLabel', {
-            defaultMessage: 'Install Beats for {solution}',
+          i18n.translate('kibana-react.noDataPage.elasticAgentCard.buttonLabel', {
+            defaultMessage: 'Find a {solution} integration',
             values: { solution },
           })}
       </EuiButton>
@@ -50,16 +52,15 @@ export const ElasticBeatsCard: FunctionComponent<ElasticBeatsCardProps> = ({
     <EuiCard
       paddingSize="l"
       href={href}
-      title={i18n.translate('kibana-react.noDataPage.elasticBeatsCard.title', {
-        defaultMessage: 'Add data with Beats',
+      title={i18n.translate('kibana-react.noDataPage.elasticAgentCard.title', {
+        defaultMessage: 'Add a {solution} integration',
+        values: { solution },
       })}
-      description={i18n.translate('kibana-react.noDataPage.elasticBeatsCard.description', {
-        defaultMessage: `
-            Beats send data from hundreds or thousands of machines and systems to Logstash or Elasticsearch.`,
+      description={i18n.translate('kibana-react.noDataPage.elasticAgentCard.description', {
+        defaultMessage: `The Elastic Agent provides a simple, unified way to
+        collect data from your machines.`,
       })}
-      image={addBasePath(
-        `${basePathUrl}elastic_beats_card_${IS_DARK_THEME ? 'dark' : 'light'}.svg`
-      )}
+      image={addBasePath(`${basePathUrl}elastic_agent_card.svg`)}
       betaBadgeLabel={recommended ? NO_DATA_RECOMMENDED : undefined}
       footer={footer}
       {...(cardRest as any)}
