+++ src/plugins/kibana_react/public/page_template/no_data_page/no_data_card/no_data_card.test.tsx
